feat(refactor): ソート実装をdomainへ統合し冗長コードを削減
