+++ src/utils/index.ts
@@ -2,5 +2,4 @@ export * from "./cardDataConverter";
 export * from "./deckCode";
 export * from "./export";
 export * from "./image";
-export * from "./sort";
 export * from "./storage";
